# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Point of Sale and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: string;
  stockQuantity: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  category: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  sku?: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  price?: number;
  /** @minLength 1 */
  category?: string;
}

export type InventoryMovementType = typeof InventoryMovementType[keyof typeof InventoryMovementType];


export const InventoryMovementType = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export interface InventoryMovement {
  id: number;
  productId: number;
  productName: string;
  type: InventoryMovementType;
  quantity: number;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type InventoryMovementInputType = typeof InventoryMovementInputType[keyof typeof InventoryMovementInputType];


export const InventoryMovementInputType = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export interface InventoryMovementInput {
  productId: number;
  type: InventoryMovementInputType;
  /** @minimum 1 */
  quantity: number;
  reason?: string;
  notes?: string;
}

export interface StockLevel {
  productId: number;
  productName: string;
  sku: string;
  category: string;
  stockQuantity: number;
  price: number;
}

export interface SaleItem {
  id: number;
  saleId: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export type SalePaymentMethod = typeof SalePaymentMethod[keyof typeof SalePaymentMethod];


export const SalePaymentMethod = {
  cash: 'cash',
  card: 'card',
  other: 'other',
} as const;

export type SaleStatus = typeof SaleStatus[keyof typeof SaleStatus];


export const SaleStatus = {
  completed: 'completed',
  voided: 'voided',
} as const;

export interface Sale {
  id: number;
  invoiceNumber: string;
  /** @nullable */
  customerName?: string | null;
  subtotal: number;
  tax: number;
  total: number;
  paymentMethod: SalePaymentMethod;
  status: SaleStatus;
  items: SaleItem[];
  createdAt: string;
}

export interface SaleItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type SaleInputPaymentMethod = typeof SaleInputPaymentMethod[keyof typeof SaleInputPaymentMethod];


export const SaleInputPaymentMethod = {
  cash: 'cash',
  card: 'card',
  other: 'other',
} as const;

export interface SaleInput {
  customerName?: string;
  paymentMethod: SaleInputPaymentMethod;
  taxRate?: number;
  /** @minItems 1 */
  items: SaleItemInput[];
}

export interface DashboardSummary {
  totalProducts: number;
  totalCategories: number;
  lowStockCount: number;
  todaySalesCount: number;
  todayRevenue: number;
  allTimeRevenue: number;
  allTimeSalesCount: number;
}

export type ListProductsParams = {
search?: string;
category?: string;
};

export type ListInventoryMovementsParams = {
productId?: number;
type?: ListInventoryMovementsType;
};

export type ListInventoryMovementsType = typeof ListInventoryMovementsType[keyof typeof ListInventoryMovementsType];


export const ListInventoryMovementsType = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export type ListLowStockParams = {
threshold?: number;
};

export type ListSalesParams = {
limit?: number;
offset?: number;
};

export type ListRecentSalesParams = {
limit?: number;
};