# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Point of Sale and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProducts: number;
  totalCategories: number;
  lowStockCount: number;
  todaySalesCount: number;
  todayRevenue: number;
  allTimeRevenue: number;
  allTimeSalesCount: number;
}