# lib/api-zod/src/generated/types/inventoryMovement.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Point of Sale and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryMovementType } from './inventoryMovementType';

export interface InventoryMovement {
  id: number;
  productId: number;
  productName: string;
  type: InventoryMovementType;
  quantity: number;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}