# lib/api-zod/src/generated/types/inventoryMovementInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Point of Sale and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryMovementInputType } from './inventoryMovementInputType';

export interface InventoryMovementInput {
  productId: number;
  type: InventoryMovementInputType;
  /** @minimum 1 */
  quantity: number;
  reason?: string;
  notes?: string;
}