# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Point of Sale and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  sku: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: string;
  stockQuantity: number;
  createdAt: Date;
  updatedAt: Date;
}