# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Point of Sale and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  sku?: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  price?: number;
  /** @minLength 1 */
  category?: string;
}